+++ dbos/core.py
@@ -1,12 +1,15 @@
 from concurrent.futures import Future
-from typing import TYPE_CHECKING, TypeVar
+from typing import TYPE_CHECKING, Any, ParamSpec, TypeVar
 
-from dbos.error import DBOSNonExistentWorkflowError
+from dbos import utils
+from dbos.error import DBOSNonExistentWorkflowError, DBOSWorkflowConflictUUIDError
+from dbos.system_database import WorkflowStatusInternal
 from dbos.workflow import WorkflowHandle, WorkflowStatus
 
 if TYPE_CHECKING:
-    from dbos.dbos import DBOS
+    from dbos.dbos import DBOS, Workflow
 
+P = ParamSpec("P")  # A generic type for workflow parameters
 R = TypeVar("R", covariant=True)  # A generic type for workflow return values
 
 
@@ -42,3 +45,29 @@ class _WorkflowHandlePolling(WorkflowHandle[R]):
         if stat is None:
             raise DBOSNonExistentWorkflowError(self.workflow_uuid)
         return stat
+
+
+def _execute_workflow(
+    dbos: "DBOS",
+    status: WorkflowStatusInternal,
+    func: "Workflow[P, R]",
+    *args: Any,
+    **kwargs: Any,
+) -> R:
+    try:
+        output = func(*args, **kwargs)
+    except DBOSWorkflowConflictUUIDError:
+        # Retrieve the workflow handle and wait for the result.
+        wf_handle: WorkflowHandle[R] = dbos.retrieve_workflow(DBOS.workflow_id)
+        output = wf_handle.get_result()
+        return output
+    except Exception as error:
+        status["status"] = "ERROR"
+        status["error"] = utils.serialize(error)
+        dbos.sys_db.update_workflow_status(status)
+        raise error
+
+    status["status"] = "SUCCESS"
+    status["output"] = utils.serialize(output)
+    dbos.sys_db.update_workflow_status(status)
+    return output
